move test type out of dto mod
